requirements.txt
python-dotenv
python-telegram-bot==3.3